pyproject.toml
[project]
name = "clients"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.6.6",
    "mcp>=1.13.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
]
